# add notes options to the instrument load