--- a/lib/index.js
+++ b/lib/index.js
@@ -30,7 +30,7 @@ Soundfont.prototype.instrument = function (name, options) {
   if (name in this.instruments) return this.instruments[name]
   var inst = { name: name, play: oscillatorPlayer(ctx, options) }
   this.instruments[name] = inst
-  var promise = loadBank(ctx, this.nameToUrl(name)).then(function (buffers) {
+  var promise = loadBank(ctx, this.nameToUrl(name), options.notes).then(function (buffers) {
     inst.play = buffersPlayer(ctx, buffers, options)
     return inst
   })