--- a/lib/index.js
+++ b/lib/index.js
@@ -11,7 +11,7 @@ var buffersPlayer = require('./buffers-player')
  * @param {Function} nameToUrl - (Optional) a function that maps the sound font name to the url
  * @return {Soundfont} a soundfont object
  */
-function Soundfont (ctx, nameToUrl) {
+function Soundfont(ctx, nameToUrl) {
   if (!(this instanceof Soundfont)) return new Soundfont(ctx)
 
   this.nameToUrl = nameToUrl || Soundfont.nameToUrl || gleitzUrl
@@ -28,14 +28,17 @@ Soundfont.prototype.instrument = function (name, options) {
   var ctx = this.ctx
   name = name || 'default'
   if (name in this.instruments) return this.instruments[name]
-  var inst = { name: name, play: oscillatorPlayer(ctx, options) }
+  var inst = {name: name, play: oscillatorPlayer(ctx, options)}
+  var notes = options ? options.notes : undefined;
   this.instruments[name] = inst
-  var promise = loadBank(ctx, this.nameToUrl(name), options.notes).then(function (buffers) {
+  var promise = loadBank(ctx, this.nameToUrl(name), notes).then(function (buffers) {
     inst.play = buffersPlayer(ctx, buffers, options)
     return inst
   })
   this.promises.push(promise)
-  inst.onready = function (cb) { promise.then(cb) }
+  inst.onready = function (cb) {
+    promise.then(cb)
+  }
   return inst
 }
 
@@ -51,7 +54,7 @@ Soundfont.loadBuffers = function (ctx, name) {
  * @param {String} name - instrument name
  * @returns {String} the Soundfont file url
  */
-function gleitzUrl (name) {
+function gleitzUrl(name) {
   return 'https://cdn.rawgit.com/gleitz/midi-js-Soundfonts/master/FluidR3_GM/' + name + '-ogg.js'
 }