--- a/lib/index.js
+++ b/lib/index.js
@@ -31,7 +31,7 @@ Soundfont.prototype.instrument = function (name, options) {
   var inst = {name: name, play: oscillatorPlayer(ctx, options)}
   var notes = options ? options.notes : undefined;
   this.instruments[name] = inst
-  var promise = loadBank(ctx, this.nameToUrl(name), notes).then(function (buffers) {
+  var promise = loadBank(ctx, this.nameToUrl(name), undefined, undefined, notes).then(function (buffers) {
     inst.play = buffersPlayer(ctx, buffers, options)
     return inst
   })