--- a/lib/load-bank.js
+++ b/lib/load-bank.js
@@ -11,18 +11,22 @@ var decodeBuffer = require('./decode-buffer')
  * @param {Function} get - (Optional) given a url return a promise with the contents
  * @param {Function} parse - (Optinal) given a js file return JSON object
  */
-module.exports = function (ctx, url, get, parse) {
-  get = get || getContent
-  parse = parse || parseJavascript
+module.exports = function (ctx, url, notes) {
+  var get = getContent
+  var parse = parseJavascript
   return Promise.resolve(url).then(get).then(parse)
     .then(function (data) {
-      return { ctx: ctx, data: data, buffers: {} }
+      return {ctx: ctx, data: data, buffers: {}}
     })
-    .then(decodeBank)
-    .then(function (bank) { return bank.buffers })
+    .then(function (data) {
+      return decodeBank(data, notes)
+    })
+    .then(function (bank) {
+      return bank.buffers
+    })
 }
 
-function getContent (url) {
+function getContent(url) {
   return new Promise(function (done, reject) {
     var req = new window.XMLHttpRequest()
     req.open('GET', url)
@@ -49,7 +53,7 @@ function getContent (url) {
  * @return {JSON} the parsed data as JSON object
  * @api private
  */
-function parseJavascript (data) {
+function parseJavascript(data) {
   var begin = data.indexOf('MIDI.Soundfont.')
   begin = data.indexOf('=', begin) + 2
   var end = data.lastIndexOf(',')
@@ -62,13 +66,24 @@ function parseJavascript (data) {
  * @return {Promise} a promise that resolves to the bank with the buffers decoded
  * @api private
  */
-function decodeBank (bank) {
+function decodeBank(bank, notes) {
   var promises = Object.keys(bank.data).map(function (note) {
-    return decodeBuffer(bank.ctx, bank.data[note])
-    .then(function (buffer) {
-      bank.buffers[midi(note)] = buffer
-    })
-  })
+    // First check is notes are passed by as param
+    if (typeof notes !== 'undefined') {
+      // if the current notes is needed for the instrument.
+      if (notes.indexOf(note) !== -1) {
+        return decodeBuffer(bank.ctx, bank.data[note])
+          .then(function (buffer) {
+            bank.buffers[midi(note)] = buffer
+          })
+      }
+    } else {
+      return decodeBuffer(bank.ctx, bank.data[note])
+        .then(function (buffer) {
+          bank.buffers[midi(note)] = buffer
+        })
+    }
+  });
 
   return Promise.all(promises).then(function () {
     return bank