--- a/lib/load-bank.js
+++ b/lib/load-bank.js
@@ -66,6 +66,7 @@ function parseJavascript(data) {
  * @return {Promise} a promise that resolves to the bank with the buffers decoded
  * @api private
  */
+
 function decodeBank(bank, notes) {
   var promises = Object.keys(bank.data).map(function (note) {
     // First check is notes are passed by as param