--- a/lib/load-bank.js
+++ b/lib/load-bank.js
@@ -63,6 +63,7 @@ function parseJavascript(data) {
 /*
  * Decode a bank
  * @param {Object} bank - the bank object
+ * @param {Array} notes - an array of required notes
  * @return {Promise} a promise that resolves to the bank with the buffers decoded
  * @api private
  */
@@ -71,8 +72,12 @@ function decodeBank(bank, notes) {
   var promises = Object.keys(bank.data).map(function (note) {
     // First check is notes are passed by as param
     if (typeof notes !== 'undefined') {
+      // convert the notes to midi number
+      var notesMidi = notes.map(function(note){
+        return midi(note);
+      })
       // if the current notes is needed for the instrument.
-      if (notes.indexOf(note) !== -1) {
+      if (notesMidi.indexOf(midi(note)) !== -1) {
         return decodeBuffer(bank.ctx, bank.data[note])
           .then(function (buffer) {
             bank.buffers[midi(note)] = buffer
@@ -90,3 +95,4 @@ function decodeBank(bank, notes) {
     return bank
   })
 }
+