--- a/lib/load-bank.js
+++ b/lib/load-bank.js
@@ -11,9 +11,9 @@ var decodeBuffer = require('./decode-buffer')
  * @param {Function} get - (Optional) given a url return a promise with the contents
  * @param {Function} parse - (Optinal) given a js file return JSON object
  */
-module.exports = function (ctx, url, notes) {
-  var get = getContent
-  var parse = parseJavascript
+module.exports = function (ctx, url, get, parse, notes) {
+  get = get || getContent
+  parse = parse || parseJavascript
   return Promise.resolve(url).then(get).then(parse)
     .then(function (data) {
       return {ctx: ctx, data: data, buffers: {}}