Hi, 

This is just a raw suggestion, when loading a instrument you load all the notes for the instruments, but sometimes you don't need all of them when playing a midi file. So i propose you can pass in an extra option for an instrument -> notes. Using this note list, the buffer will only load the given notes and not everything in the file.

This is mainly a performance update, because you are base64 decoding and mp3 decoding all the notes. What can be very cpu hungry when loading instruments. In my application i could bring down load time significantly:

I added a timer start before instrument loades and a timer stop on soundfont.onready

For example a midifile with 16 different instruments(all in browser cache):
before: 11758 ms
after: 2253 ms
For example a midifile with 1 instrument(in browser cache):
before: 739 ms
after: 148 ms


Again this is a very raw pull request, any suggestions or style remarks are welcome.