@danigb 2016-05-11
Hi @frankbaele this is a nice one!! Thanks. I will review ASAP and merge if everything looks ok to me 👍  Thanks again!