@danigb 2016-05-11
Hi @frankbaele,

As I can see, you remove two parameters from `loadBank` function. They are not used by production code, but they are used to test the function. I agree that's not very elegant solution, and I'm open to new approaches, but I think that If I merge you changes, the test will break. Have you tried to run the tests after your changes? 

Cheers,
Dani