@danigb 2016-05-11
Hi, I've cloned your repo and saw the test failing. I'm not very used to github collaboration, so I don't know how to proceed... Merge your changes and fix the tests? You do it in your repo? I clone your repo and pull request to your repo? :-0 ... Any advice?