@danigb 2016-05-11
By the way, I like a lot your proposal. Let's put it inside the library!