@frankbaele 2016-05-11
i will fix the tests tommorow morning and commit them to my repo, this will
update the pull request.

On Wed, May 11, 2016 at 8:54 PM, danigb <notifications@github.com> wrote:

> By the way, I like a lot your proposal. Let's put it inside the library!
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly or view it on GitHub
> <https://github.com/danigb/soundfont-player/pull/22#issuecomment-218555262>
>