@frankbaele 2016-05-12
Ok, fixed the tests quickly but, there are better ways to do your depency injection for testing. If i find time the comming weeks, i will post an example.