@danigb 2016-05-12
Hi @frankbaele 

Finally I made some change to modify the dependency injection setup. Anyway, suggestions and PR are welcomed. I've published your changes into 0.8.2 npm version, and I've added eslint and travis-ci support.

Thanks again for your PR! 
Regards,
Dani